# ezkey-demo-app-acme/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ezkey Admin API",
    "description": "Administration API for Ezkey - Open Source Cryptographic MFA Platform\n\nThis API enables administrative management of Ezkey's main entities:\n- **Integrations**: Applications or systems protected by MFA\n- **Enrollments**: Associations between users, devices and integrations\n- **Auth Attempts**: MFA authentication attempts\n- **Admin Management**: Administrator authentication,\n                        enrollment recovery, and admin operations\n\nEzkey is intentionally distinct from FIDO2/WebAuthn and follows its own\ncryptographic MFA model. The API uses REST conventions and DTOs for all\nrequests and responses.\n",
    "contact": {
      "name": "Ezkey Team",
      "url": "https://ezkey.org",
      "email": "contributors@ezkey.org"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Development server Admin API"
    },
    {
      "url": "https://admin-api.ezkey.org",
      "description": "Production server Admin API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/auth/passwordless-wait": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Wait for passwordless authentication device response",
        "description": "Polls for passwordless authentication completion in two-call flows. Blocks for up to the wait window aligned with auth attempt TTL (capped at 300s). Supports both challenge-based (with display code) and non-blocking flows. Errors follow RFC 9457 Problem Details format.",
        "operationId": "passwordlessWait",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPasswordlessWaitRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authentication successful. Returns bearer token and admin session details.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Device denied, signature failed, or challenge invalid. Returns RFC 9457 ProblemDetail with specific error type URI.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid challenge code. Returns RFC 9457 ProblemDetail: type='...invalid-credentials'",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout - No device response within the wait window (max 300 seconds). Returns RFC 9457 ProblemDetail: type='...auth-timeout'",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "Admin Authentication"
        ],
        "summary": "Authenticate administrator with passwordless Ezkey MFA",
        "description": "Passwordless authentication for administrators using device-based cryptography. Supports single-call (blocking) or two-call (challenge) flows. Returns bearer token on success, or challenge info for two-step authentication. Rate limiting applied per IP. Errors follow RFC 9457 Problem Details format.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authentication successful or pending. If status='approved', bearer token is present. If status='pending', authAttemptId and challengeCode present for two-call flow.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Authentication failed with temporal issues (expired, rejected, invalid signature). Returns RFC 9457 ProblemDetail with specific error type URI.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials (username not found). Returns RFC 9457 ProblemDetail: type='...invalid-credentials'",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Account inactive or no device enrolled. Returns RFC 9457 ProblemDetail: type='...account-inactive' or '...no-enrollment'",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "408": {
            "description": "Request Timeout - No device response within the wait window (aligned with auth attempt TTL, maximum 300 seconds). Returns RFC 9457 ProblemDetail: type='...auth-timeout'",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded for this IP address. Applies after 10 failed attempts in 15 minutes.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminPasswordlessWaitRequestDto": {
        "type": "object",
        "properties": {
          "authAttemptId": {
            "type": "integer",
            "format": "int32"
          },
          "challengeCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "authAttemptId"
        ]
      },
      "AdminLoginResponseDto": {
        "type": "object",
        "description": "Response DTO for passwordless administrator login",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the authentication was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Response message describing authentication result",
            "example": "Authentication successful"
          },
          "status": {
            "type": "string",
            "description": "Authentication attempt status",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "example": "pending"
          },
          "token": {
            "type": "string",
            "description": "Bearer token for authenticated API requests",
            "example": "ezkey_abc123def456..."
          },
          "adminType": {
            "type": "string",
            "description": "Type of administrator",
            "enum": [
              "GLOBAL_ADMIN",
              "TENANT_ADMIN",
              "INTEGRATION_ADMIN"
            ],
            "example": "GLOBAL_ADMIN"
          },
          "username": {
            "type": "string",
            "description": "Administrator username",
            "example": "admin"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Expiration timestamp (UTC). When authentication succeeded: bearer token expiry. When status is pending (two-step passwordless): authentication attempt expiry — matches the persisted attempt and core setting ezkey.core.auth-attempt.ttl-seconds (not a fixed duration).",
            "example": "2025-10-15T14:30:00+01:00"
          },
          "authAttemptId": {
            "type": "integer",
            "format": "int32",
            "description": "Authentication attempt ID for two-step flow",
            "example": 123
          },
          "challengeCode": {
            "type": "integer",
            "format": "int32",
            "description": "6-digit challenge code for device verification",
            "example": 654321
          }
        }
      },
      "AdminLoginRequestDto": {
        "type": "object",
        "description": "Request DTO for passwordless administrator login",
        "properties": {
          "username": {
            "type": "string",
            "description": "Administrator username for passwordless authentication",
            "example": "admin",
            "maxLength": 50,
            "minLength": 3
          },
          "challengeRequested": {
            "type": "boolean",
            "description": "Request challenge verification on device (6-digit code). When true, returns authAttemptId and challengeCode for two-step flow",
            "example": false
          },
          "nonBlocking": {
            "type": "boolean",
            "description": "Request immediate response with authAttemptId and expiresAt instead of blocking until device responds. Allows client to display countdown timer and poll /passwordless-wait endpoint. When false (default), blocking wait is used when no challenge is required (backward compatible). This flag has no effect when challenge is required (challenge flow is always non-blocking).",
            "example": false
          }
        },
        "required": [
          "username"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT authentication token for admin API",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}